sourcegraph: Trace indexing
This is the start of adding more features to our indexserver. Initially we
expose the traces on each request. This should be a much nicer way to inspect
problems via port-forwarding the web interface. As we improve, we will likely
show more of this information in a nicer way in the main sourcegraph UI.