indexserver: Index repositories concurrently
gitserver is IO bound. So more than likely streaming the archive is slower
than the time it takes to index a repo on all cores. As such it is better to
have several archives stream at once.

There are further perf improvements here that can be done if this is
insufficient for our large customers:

* Call resolve revision with high concurrency.
* Track revision across runs. Prioritize unseen revisions.